Updated formatting

+++ file_converter.sh
@@ -8,9 +8,7 @@ while true; do
 	readarray -d '' -O "${#videos[@]}" videos < <(find . -not -path "*/.*" -type f -name "*".$format -print0)
 
 	read -p "Do you want to add another file format to convert? (Y/N) " choice
-	if [ $choice == "N" || $choice == "n" ]; then
-		break
-	fi
+	if [ $choice == "N" || $choice == "n" ]; then break; fi
 done
 
 # Set arguments for ffmpeg video conversion
@@ -31,7 +29,5 @@ for (( i=0; i<$len; i++ )); do
 
 	# Prompt user wether to delete an original video file
 	read -p "Do you want to delete an original file? (Y/N)" choice
-	if [ $choice == "Y" ]; then
-		rm ${videos[$i]}
-	fi
+	if [ $choice == "Y" ]; then rm ${videos[$i]}; fi
 done
